redesign the project structure
